refactor: Move db package to augmented package
