+++ src/rago/augmented/db/__init__.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-from rago.db.base import DBBase
-from rago.db.faiss import FaissDB
+from rago.augmented.db.base import DBBase
+from rago.augmented.db.faiss import FaissDB
 
 __all__ = [
     'DBBase',
