+++ src/rago/augmented/db/base.py
